+++ internal/server/v2/observation/direct.go
@@ -23,9 +23,9 @@ import (
 	"sort"
 	"strings"
 
+	"github.com/datacommonsorg/mixer/internal/merger"
 	pb "github.com/datacommonsorg/mixer/internal/proto"
 	pbv2 "github.com/datacommonsorg/mixer/internal/proto/v2"
-	"github.com/datacommonsorg/mixer/internal/response_merger.go"
 	"github.com/datacommonsorg/mixer/internal/server/ranking"
 	"github.com/datacommonsorg/mixer/internal/server/stat"
 	"github.com/datacommonsorg/mixer/internal/store"
@@ -104,7 +104,7 @@ func FetchDirect(
 		return nil, err
 	}
 	// Prefer SQL data over BT data, so put sqlObservation first.
-	return response_merger.MergeObservation(sqlObservation, btObservation), nil
+	return merger.MergeObservation(sqlObservation, btObservation), nil
 }
 
 // FetchDirectBT fetches data from Bigtable cache.
